requirements.txt
aiogram
ollama
aiohttp
torch
python-dotenv
